refactor: mark pythonnet internal
