+++ src/data_factory_testing_framework/_functions/functions_date_implementation.py
@@ -1,4 +1,4 @@
-from data_factory_testing_framework.pythonnet.csharp_datetime import CSharpDateTime
+from data_factory_testing_framework._pythonnet.csharp_datetime import CSharpDateTime
 
 
 def utcnow(fmt: str = None) -> str:
