+++ tests/unit/functions/test_functions_date_functions.py
@@ -1,7 +1,7 @@
 import data_factory_testing_framework._functions.functions_date_implementation as date_functions
 import pytest
 from _pytest.mark import param
-from data_factory_testing_framework.pythonnet.csharp_datetime import CSharpDateTime
+from data_factory_testing_framework._pythonnet.csharp_datetime import CSharpDateTime
 
 
 def test_utcnow(monkeypatch: pytest.MonkeyPatch) -> None:
